// src/@saas-boilerplate/providers/payment/types.ts
export type CyclePeriod = 'day' | 'week' | 'month' | 'year'

/**
 * Represents the status of a subscription.
 * @enum {string}
 */
export type SubscriptionStatus =
  | 'active'
  | 'canceled'
  | 'incomplete'
  | 'incomplete_expired'
  | 'past_due'
  | 'unpaid'
  | 'trialing'
// ...
export type Customer = {
  id: string
  providerId: string
  organizationId: string
  name: string
  email: string
  metadata?: Record<string, unknown>

  subscription?: {
    id: string
    providerId: string
    status: SubscriptionStatus
    trialDays: number | null

    plan: {
      id: string
      providerId: string
      slug: string
      name: string
      description: string
      metadata: PlanMetadata
      price: {
        id: string
        providerId: string
        slug: string
        amount: number
        currency: string
        interval: CyclePeriod
      }
    }

    usage: Usage[]

    createdAt?: Date
    updatedAt?: Date
  }

  createdAt?: Date
  updatedAt?: Date
}
// ...
export type PlanFeature = {
  slug: string
  name: string
  enabled: boolean
  description?: string
  table?: string
  limit?: number
  cycle?: CyclePeriod
}

/**
 * Represents metadata associated with a subscription plan.
 * @interface PlanMetadata
 * @property {PlanFeature[]} features - List of features included in this plan
 */
export type PlanMetadata = {
  features: PlanFeature[]
}
// ...
export type Plan = {
  id: string
  providerId: string
  slug: string
  name: string
  description: string
  metadata: PlanMetadata
  prices: Price[]
  archived?: boolean
  createdAt?: Date
  updatedAt?: Date
}
// ...
export type Price = {
  id: string
  providerId: string
  planId: string
  slug: string
  amount: number
  currency: string
  interval: CyclePeriod
  intervalCount: number
  metadata?: Record<string, unknown>
  active?: boolean
  createdAt?: Date
  updatedAt?: Date
}
// ...
export type Subscription = {
  id: string
  providerId: string
  customerId: string
  priceId: string
  quantity?: number
  trialDays?: number
  metadata?: Record<string, unknown>
  billingCycleAnchor?: Date
  prorationBehavior?: 'create_prorations' | 'none'
  status: SubscriptionStatus
  createdAt?: Date
  updatedAt?: Date
}
// ...
export type Usage = {
  slug: string
  name: string
  description?: string
  usage: number
  limit: number
  lastReset?: Date
  nextReset?: Date
  cycle: CyclePeriod
}
// ...
export type CustomerDTO = {
  providerId?: string
  referenceId: string
  name: string
  email: string
  metadata?: Record<string, unknown>
}
// ...
export type PriceDTO = {
  providerId?: string
  planId: string
  slug: string
  amount: number
  currency: string
  interval: CyclePeriod
  intervalCount: number
  metadata?: Record<string, unknown>
}
// ...
export type PlanDTO = {
  providerId?: string
  slug: string
  name: string
  description: string
  metadata: PlanMetadata
  prices: Array<{
    providerId?: string
    slug: string
    interval: CyclePeriod
    amount: number
    currency: string
    intervalCount: number
    metadata?: Record<string, unknown>
  }>
}
// ...
export type SubscriptionDTO = {
  status: SubscriptionStatus
  providerId?: string
  customerId: string
  priceId: string
  quantity?: number
  trialDays?: number
  billingCycleAnchor?: Date
  prorationBehavior?: 'create_prorations' | 'none'
  metadata?: Record<string, unknown>
}